requirements.txt
streamlit
ultralytics==8.0.173
opencv-python-headless
Pillow
numpy
gdown
torch